Split test_base in two separate test files.
